Resolve race between cloudtrail and s3 bucket
We require the bucket to be created and have the correct policy attached
to it otherwise we receive InsufficientS3BucketPolicyException at
cloudtrail creation time. To do this, we extract policy
attachment to separate resource to resolve cycle of
the policy including the ARN of the bucket and we
add a dependency on the policy attachment to the cloudtrail resource.
This ensures proper ordering of resource creation.